Version 0.0.1
This version was created by WinPEBuilder from oficial Microsoft Windows 7 install disk and with WinPE7SE and VistaPE projects.
Some source code of scripts from projects was changed, they can be downloaded from http://yadi.sk/d/8UUdzJ5a8GAqW
This version is pre-alpha, so there are lots of bugs, size of LiveCD is very large and it needs ~800 Mb of RAM.
There sre a lot of programm that does not works or uses a lot of RAM.

So, I not recommend to use this version at all. But if the first version that really works.